doc(inputNumber): hide the addonAfter and addonBefore properties
