--- components/input-number/demo/addon.md
@@ -1,7 +1,0 @@
-## zh-CN
-
-用于配置一些固定组合。
-
-## en-US
-
-Using pre & post tabs example.
