+++ components/input-number/demo/compact.md
@@ -1,0 +1,7 @@
+## zh-CN
+
+使用 Space.Compact 替代 `addonAfter` 和 `addonBefore`。
+
+## en-US
+
+Use Space.Compact instead of `addonAfter` and `addonBefore`.
